LayoutTests/media/modern-media-controls/tracks-button/tracks-button-expected.txt

Testing the TracksButton class.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS tracksButton.element.localName is "button"
PASS tracksButton.element.classList.contains("icon") is true
PASS tracksButton.element.classList.contains("tracks") is true
PASS tracksButton.iconName is Icons.Tracks
PASS tracksButton.element.style.webkitMaskImage.includes("macOS/media-selection@") became true
PASS fullscreenTracksButton.element.style.webkitMaskImage.includes("macOS/media-selection-fullscreen@") became true
PASS successfullyParsed is true

TEST COMPLETE
